Combine python scripts into one file
